+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.18] - 2025-06-12
+### Added
+* property `NumberOfNeighbors` to `Node`
+* function `RemoveNeighbor` to `Node`
+
 ## [0.0.17] - 2025-06-03
 ### Changed
 * all file headers and file dividers to match new Kokowolo standards